fix circular dependencies in product-list
